rename and restructure the ItemPages
