+++ TkMobile/TkMobile/ItemPages/LocomotiveOverviewTile.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="TkMobile.ItemOverviews.LocomotiveOverviewTile">
+             x:Class="TkMobile.ItemPages.LocomotiveOverviewTile">
     <ContentPage.Content>
         <AbsoluteLayout x:Name="ContentLayout"/>
     </ContentPage.Content>
